fix API implementation & removed video handling
YouTube API can parse 50 id's per call and the cost is 5 for the entire operation, rather than 5 for each song. This can be further optimized by only calling for titles where necessary, resulting in 3 units for most calls to just get a duration or 5 for calls with titles. Might look into that later. For any videos which the API cannot find for one reason or another, they are changed to Unknown Artist - Unavailable Video rather than being left as parsme - parseme.